Check user access to the resource

+++ src/main/java/ru/hse/rekoder/controllers/FolderController.java
@@ -2,7 +2,9 @@ package ru.hse.rekoder.controllers;
 
 import org.springframework.beans.BeanUtils;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
+import ru.hse.rekoder.exceptions.AccessDeniedException;
 import ru.hse.rekoder.model.Folder;
 import ru.hse.rekoder.requests.FolderRequest;
 import ru.hse.rekoder.requests.ProblemIdWrap;
@@ -10,6 +12,7 @@ import ru.hse.rekoder.responses.FolderResponse;
 import ru.hse.rekoder.responses.ProblemResponse;
 import ru.hse.rekoder.services.FolderService;
 import ru.hse.rekoder.services.JsonMergePatchService;
+import ru.hse.rekoder.services.TeamService;
 
 import javax.json.JsonMergePatch;
 import javax.validation.Valid;
@@ -22,11 +25,14 @@ import java.util.stream.Collectors;
 public class FolderController {
     private final FolderService folderService;
     private final JsonMergePatchService jsonMergePatchService;
+    private final TeamService teamService;
 
     public FolderController(FolderService folderService,
-                            JsonMergePatchService jsonMergePatchService) {
+                            JsonMergePatchService jsonMergePatchService,
+                            TeamService teamService) {
         this.folderService = folderService;
         this.jsonMergePatchService = jsonMergePatchService;
+        this.teamService = teamService;
     }
 
     @GetMapping("/{folderId}")
@@ -37,7 +43,9 @@ public class FolderController {
 
     @PostMapping("/{folderId}/folders")
     public ResponseEntity<FolderResponse> createFolder(@PathVariable int folderId,
-                                                       @RequestBody @Valid FolderRequest folderRequest) {
+                                                       @RequestBody @Valid FolderRequest folderRequest,
+                                                       Authentication authentication) {
+        checkAccess(folderId, authentication.getName());
         Folder folder = new Folder();
         BeanUtils.copyProperties(folderRequest, folder);
         Folder createdFolder = folderService.createNewFolder(folderId, folder);
@@ -61,20 +69,26 @@ public class FolderController {
     }
 
     @PatchMapping("/{folderId}/problems")
-    public ResponseEntity<?> addProblemToFolder(@PathVariable int folderId, @RequestBody ProblemIdWrap problemToAdd) {
+    public ResponseEntity<?> addProblemToFolder(@PathVariable int folderId, @RequestBody ProblemIdWrap problemToAdd,
+                                                Authentication authentication) {
+        checkAccess(folderId, authentication.getName());
         folderService.addProblemToFolder(folderId, problemToAdd.getProblemId());
         return ResponseEntity.noContent().build();
     }
 
     @DeleteMapping("/{folderId}/problems/{problemId}")
-    public ResponseEntity<?> deleteProblemFromFolder(@PathVariable int folderId, @PathVariable int problemId) {
+    public ResponseEntity<?> deleteProblemFromFolder(@PathVariable int folderId, @PathVariable int problemId,
+                                                     Authentication authentication) {
+        checkAccess(folderId, authentication.getName());
         folderService.deleteProblemFromFolder(folderId, problemId);
         return ResponseEntity.noContent().build();
     }
 
     @PatchMapping(path = "/{folderId}", consumes = "application/merge-patch+json")
     public ResponseEntity<FolderResponse> updateProblem(@PathVariable int folderId,
-                                                         @Valid @RequestBody JsonMergePatch jsonMergePatch) {
+                                                         @Valid @RequestBody JsonMergePatch jsonMergePatch,
+                                                        Authentication authentication) {
+        checkAccess(folderId, authentication.getName());
         Folder folder = folderService.getFolder(folderId);
         BeanUtils.copyProperties(
                 jsonMergePatchService.mergePatch(jsonMergePatch, convertToRequest(folder), FolderRequest.class),
@@ -84,9 +98,23 @@ public class FolderController {
     }
 
     @DeleteMapping("/{folderId}")
-    public ResponseEntity<?> deleteFolder(@PathVariable int folderId) {
+    public ResponseEntity<?> deleteFolder(@PathVariable int folderId,
+                                          Authentication authentication) {
+        checkAccess(folderId, authentication.getName());
         folderService.deleteFolder(folderId);
         return ResponseEntity.noContent().build();
+    }
+
+    private void checkAccess(int folderId, String username) {
+        Folder folder = folderService.getFolder(folderId);
+        if (folder.getOwnerId().getProblemOwnerType().equals("user")
+                && !folder.getOwnerId().getProblemOwnerId().equals(username)) {
+            throw new AccessDeniedException();
+        } else if (folder.getOwnerId().getProblemOwnerType().equals("team")) {
+            if (!teamService.getTeam(folder.getOwnerId().getProblemOwnerId()).getMemberIds().contains(username)) {
+                throw new AccessDeniedException();
+            }
+        }
     }
 
     private FolderRequest convertToRequest(Folder folder) {
